[refactor] Refactor example/issues files
