+++ example/stories/issues/59/child.vue
